requirements.txt
tensorflow>=2.10.0
datasets>=2.0.0
Pillow>=9.0.0
numpy>=1.21.0
huggingface-hub>=0.10.0
